fix: adicionando validacao da API

+++ app/Http/Controllers/MarcaController.php
@@ -55,6 +55,10 @@ class MarcaController extends Controller
     public function show($id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null)
+        {
+            return ['erro' => 'Recurso Pesquisado não existe'];
+        }
         return $marca;
     }
 
@@ -80,10 +84,14 @@ class MarcaController extends Controller
     {
         //$marca->update($request->all());
         $marca = $this->marca->find($id);
+        if ($marca === null)
+        {
+            return ['erro' => 'Impossível realizar a atualização. O recurso solicitado não existe'];
+        }
         $marca->update($request->all());
         return $marca;
     }
-
+    
     /**
      * Remove the specified resource from storage.
      *
@@ -93,6 +101,10 @@ class MarcaController extends Controller
     public function destroy($id)
     {
         $marca = $this->marca->find($id);
+        if ($marca === null)
+        {
+            return ['erro' => 'Impossível realizar a exclusão. O recurso solicitado não existe'];
+        }
         $marca->delete();
         return ['msg' => 'A marca foi removida com sucesso!'];
     }
